split insert.py in multiple files
